+++ Repository/IInscricaoRepository.cs
@@ -1,0 +1,9 @@
+using CarrefourApi.Model;
+
+namespace CarrefourApi.Repository
+{
+    public interface IInscricaoRepository
+    {
+        void InserirInscricao(Inscricao inscricao);
+    }
+}
