+++ Repository/MongoRepository/MongoInscricaoRepository.cs
@@ -1,0 +1,13 @@
+
+using CarrefourApi.Model;
+
+namespace CarrefourApi.Repository.MongoRepository
+{
+    public class MongoInscricaoRepository : IInscricaoRepository
+    {
+        public void InserirInscricao(Inscricao inscricao)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
